Extract exceptions to SchemaUtils
