+++ json-schema-core/src/test/resources/jsonschema/3_true_schema.json
